+++ flag/service/collector/collector.go
@@ -1,7 +1,11 @@
 package collector
 
-import "github.com/giantswarm/app-exporter/flag/service/collector/apps"
+import (
+	"github.com/giantswarm/app-exporter/flag/service/collector/apps"
+	"github.com/giantswarm/app-exporter/flag/service/collector/provider"
+)
 
 type Collector struct {
-	Apps apps.Apps
+	Apps     apps.Apps
+	Provider provider.Provider
 }