+++ integration/templates/template.go
@@ -1,0 +1,11 @@
+// +build k8srequired
+
+package templates
+
+// AppExporterValues values required by app-exporter chart.
+const AppExporterValues = `Installation:
+  V1:
+    Provider:
+      Kind: aws
+    Registry:
+      Domain: quay.io`